requirements.txt
openai==1.7.2
Pillow==10.2.0
pytesseract==0.3.10
streamlit==1.30.0
streamlit_option_menu==0.3.6
